telemetry: Make prom metrics utils usable throughout the server
This commit is pure code movement, so that we can use prometheus metrics to
report on general materialized usage.
