+++ src/materialized/src/http/metrics.rs
@@ -9,70 +9,16 @@
 
 //! Metrics HTTP endpoints.
 
+use crate::BUILD_INFO;
 use std::collections::{BTreeMap, BTreeSet};
-use std::fmt;
 use std::time::Instant;
 
 use askama::Template;
 use hyper::{Body, Request, Response};
-use lazy_static::lazy_static;
-use prometheus::{
-    register_gauge_vec, register_int_gauge_vec, Encoder, Gauge, GaugeVec, IntGauge, IntGaugeVec,
-};
-use sysinfo::{ProcessorExt, SystemExt};
+use prometheus::Encoder;
 
 use crate::http::util;
-use crate::BUILD_INFO;
-
-lazy_static! {
-    static ref SERVER_METADATA_RAW: GaugeVec = register_gauge_vec!(
-        "mz_server_metadata_seconds",
-        "server metadata, value is uptime",
-        &[
-            "build_time",
-            "build_version",
-            "build_sha",
-            "os",
-            "ncpus_logical",
-            "ncpus_physical",
-            "cpu0",
-            "memory_total",
-        ]
-    )
-    .expect("can build mz_server_metadata");
-    static ref SERVER_METADATA: Gauge = {
-        let mut system = sysinfo::System::new();
-        system.refresh_system();
-
-        SERVER_METADATA_RAW.with_label_values(&[
-            BUILD_INFO.time,
-            BUILD_INFO.version,
-            BUILD_INFO.sha,
-            &os_info::get().to_string(),
-            &num_cpus::get().to_string(),
-            &num_cpus::get_physical().to_string(),
-            &{
-                let cpu0 = &system.get_processors()[0];
-                format!("{} {}MHz", cpu0.get_brand(), cpu0.get_frequency())
-            },
-            &system.get_total_memory().to_string(),
-        ])
-    };
-    pub static ref WORKER_COUNT: IntGaugeVec = register_int_gauge_vec!(
-        "mz_server_metadata_timely_worker_threads",
-        "number of timely workers materialized is running with",
-        &["count"]
-    )
-    .unwrap();
-    static ref REQUEST_TIMES: IntGaugeVec = register_int_gauge_vec!(
-        "mz_server_scrape_metrics_times",
-        "how long it took to gather metrics, used for very low frequency high accuracy measures",
-        &["action"]
-    )
-    .unwrap();
-    static ref REQUEST_METRICS_GATHER: IntGauge = REQUEST_TIMES.with_label_values(&["gather"]);
-    static ref REQUEST_METRICS_ENCODE: IntGauge = REQUEST_TIMES.with_label_values(&["encode"]);
-}
+use crate::server_metrics::{load_prom_metrics, PromMetric, REQUEST_METRICS_ENCODE};
 
 #[derive(Template)]
 #[template(path = "http/templates/status.html")]
@@ -89,9 +35,8 @@ pub async fn handle_prometheus(
     start_time: Instant,
 ) -> Result<Response<Body>, anyhow::Error> {
     let metric_families = load_prom_metrics(start_time);
-    let encoder = prometheus::TextEncoder::new();
     let mut buffer = Vec::new();
-
+    let encoder = prometheus::TextEncoder::new();
     let start = Instant::now();
     encoder.encode(&metric_families, &mut buffer)?;
     REQUEST_METRICS_ENCODE.set(Instant::now().duration_since(start).as_micros() as i64);
@@ -175,107 +120,4 @@ pub async fn handle_status(
         start_time,
         metrics: metrics.values().collect(),
     }))
-}
-
-/// Call [`prometheus::gather`], ensuring that all our metrics are up to date
-fn load_prom_metrics(start_time: Instant) -> Vec<prometheus::proto::MetricFamily> {
-    let before_gather = Instant::now();
-    let uptime = before_gather - start_time;
-    let (secs, milli_part) = (uptime.as_secs() as f64, uptime.subsec_millis() as f64);
-    SERVER_METADATA.set(secs + milli_part / 1_000.0);
-    let result = prometheus::gather();
-
-    REQUEST_METRICS_GATHER.set(Instant::now().duration_since(before_gather).as_micros() as i64);
-    result
-}
-
-#[derive(Debug)]
-enum PromMetric<'a> {
-    Counter {
-        name: &'a str,
-        value: f64,
-        labels: BTreeMap<&'a str, &'a str>,
-    },
-    Gauge {
-        name: &'a str,
-        value: f64,
-        labels: BTreeMap<&'a str, &'a str>,
-    },
-    Histogram {
-        name: &'a str,
-        count: u64,
-        labels: BTreeMap<&'a str, &'a str>,
-    },
-}
-
-impl fmt::Display for PromMetric<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        fn fmt(
-            f: &mut fmt::Formatter,
-            name: &str,
-            value: impl fmt::Display,
-            labels: &BTreeMap<&str, &str>,
-        ) -> fmt::Result {
-            write!(f, "{} ", name)?;
-            for (n, v) in labels.iter() {
-                write!(f, "{}={} ", n, v)?;
-            }
-            writeln!(f, "{}", value)
-        }
-        match self {
-            PromMetric::Counter {
-                name,
-                value,
-                labels,
-            } => fmt(f, name, value, labels),
-            PromMetric::Gauge {
-                name,
-                value,
-                labels,
-            } => fmt(f, name, value, labels),
-            PromMetric::Histogram {
-                name,
-                count,
-                labels,
-            } => fmt(f, name, count, labels),
-        }
-    }
-}
-
-impl PromMetric<'_> {
-    fn from_metric_family<'a>(
-        m: &'a prometheus::proto::MetricFamily,
-    ) -> Result<Vec<PromMetric<'a>>, ()> {
-        use prometheus::proto::MetricType;
-        fn l2m(metric: &prometheus::proto::Metric) -> BTreeMap<&str, &str> {
-            metric
-                .get_label()
-                .iter()
-                .map(|lp| (lp.get_name(), lp.get_value()))
-                .collect()
-        }
-        m.get_metric()
-            .iter()
-            .map(|metric| {
-                Ok(match m.get_field_type() {
-                    MetricType::COUNTER => PromMetric::Counter {
-                        name: m.get_name(),
-                        value: metric.get_counter().get_value(),
-                        labels: l2m(metric),
-                    },
-                    MetricType::GAUGE => PromMetric::Gauge {
-                        name: m.get_name(),
-                        value: metric.get_gauge().get_value(),
-                        labels: l2m(metric),
-                    },
-                    MetricType::HISTOGRAM => PromMetric::Histogram {
-                        name: m.get_name(),
-                        count: metric.get_histogram().get_sample_count(),
-                        labels: l2m(metric),
-                    },
-                    _ => return Err(()),
-                })
-            })
-            .collect()
-    }
 }
